mysite/requirements.txt
asgiref==3.5.1
Django==4.0.4
django-debug-toolbar==3.4.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.13.1
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
